refactor bllom filter
(cherry picked from commit 05df2b7232d8b9729115d96bd9d8d6e2927ae30d)

+++ src/main/java/io/github/mtrevisan/hunlinter/datastructures/bloomfilter/hash/CRC32HashFunction.java
@@ -39,10 +39,10 @@ public class CRC32HashFunction implements HashFunction{
 	}
 
 	@Override
-	public final long hash(final byte[] bytes){
+	public final int hash(final byte[] bytes){
 		final Checksum crc32 = new CRC32();
 		crc32.update(bytes);
-		return crc32.getValue();
+		return (int)crc32.getValue();
 	}
 
 	@Override
